fix: properly set theme variables from partners
