+++ src/contexts/ThemeProvider/test-utils.tsx
@@ -1,14 +1,15 @@
 import { render, RenderOptions } from '@testing-library/react'
 import React, { ReactElement } from 'react'
 import { I18nextProvider, useTranslation } from 'react-i18next'
-import { defaultTheme } from './DefaultTheme'
+import { createTheme } from './createTheme'
 import { ThemeProvider } from './ThemeProvider'
 import '@/i18n'
 
 const AllTheProviders = ({ children }: { children: React.ReactNode }) => {
   const { i18n } = useTranslation()
+  const theme = createTheme()
   return (
-    <ThemeProvider theme={defaultTheme}>
+    <ThemeProvider theme={theme}>
       <I18nextProvider i18n={i18n}>{children}</I18nextProvider>
     </ThemeProvider>
   )
